Switched the VAE to use the generic super class.
