+++ PanoramAI/VAEorama.py
@@ -4,10 +4,7 @@ import numpy as np
 import time
 
 import tensorflow as tf
-import tensorflow.keras.layers as tfkl
-from tensorflow.keras.callbacks import EarlyStopping
 from tensorflow.keras.models import Model, Sequential
-from tensorflow.keras.initializers import RandomNormal
 
 class VAEorama(GENERICorama):
     """Variational autoencoder (VAE) used to learn panoramic images.
@@ -35,94 +32,49 @@ class VAEorama(GENERICorama):
                  latent_dim = 100):
         super().__init__(dataset, BATCH_SIZE, test_size, latent_dim)
 
+    def _generate_random_vector(self, n_samples):
+        self.n_samples_to_generate = n_samples
+        self.random_vector_for_generation = tf.random.normal(
+            shape=[n_samples, self.latent_dim])
+        return
 
-    def _log_normal_pdf(self, sample, mean, logvar, raxis=1):
-        """Log PDF of the multivariate normal distribution
-        given tensors of means and log(variances). Note that
-        the element-wise log-pdf is reduced along `raxis`.
+    def generate_samples(self, n_samples):
+        if n_samples > self.n_samples_to_generate:
+            print("Regenerating sample vector.")
+            self._generate_random_vector(n_samples)
+        return self.model.sample(self.random_vector_for_generation)
 
-        Args:
-            sample (`tensor`): random variable
-            mean (`tensor`): mean of distributions
-            logvar (`tensor`): log of the variances
-            raxis (int): axis to take the sum over; default 1
-        
-        Returns:
-            (`float`): log pdf of the sample
+    def create_model(self):
+        M, N = self.dimensions
+        self.model = _CVAE(M, N, self.latent_dim)
+        return
 
-        """
+    def log_normal_pdf(self, sample, mean, logvar, raxis=1):
         log2pi = tf.math.log(2. * np.pi)
         return tf.reduce_sum(
             -.5 * ((sample - mean) ** 2. * tf.exp(-logvar) + logvar + log2pi),
             axis=raxis)
 
-    """
-    def VAE_loss(self):
-        # KL loss - how to pass in x???
-        mean, logvar = tf.split(self.encoder(x), 
-                                num_or_size_splits=2, axis=1)
-        #Sample from a multivariate Gaussian
-        z = tf.random.normal(shape=mean.shape) * tf.exp(logvar * .5) + mean
-        logpz = self._log_normal_pdf(z, 0., 0.)
-        logqz_x = self._log_normal_pdf(z, mean, logvar)
+    @tf.function
+    def compute_loss(self, x):
+        mean, logvar = self.model.encode(x)
+        z = self.model.reparameterize(mean, logvar)
+        x_predicted = self.model.decode(z)
+        MSE = tf.losses.MSE(x, x_predicted)
+        logpx_z = -tf.reduce_sum(MSE)
+        logpz = self.log_normal_pdf(z, 0., 0.)
+        logqz_x = self.log_normal_pdf(z, mean, logvar)
+        return -tf.reduce_mean(logpx_z + logpz - logqz_x)
 
-        def total_loss(y_true, y_pred):
-            # Reconstruction loss
-            MSE = tf.losses.MSE(y_true, y_pred)
-            logpx_z = -tf.reduce_sum(MSE)
-            return -tf.reduce_mean(logpx_z + logpz - logqz_x)
+    @tf.function
+    def compute_apply_gradients(self, x):
+        with tf.GradientTape() as tape:
+            loss = self.compute_loss(x)
+            gradients = tape.gradient(loss, self.model.trainable_variables)
+            self.optimizer.apply_gradients(
+                zip(gradients, self.model.trainable_variables))
 
-        return total_loss
-        """
-
-    def create_model(self):
-        """Create the convolutional variation autoencoder
-        model. This function is called automatically by the 
-        constructor of the superclass.
-
-        """
-        M, N = self.dimensions
-        latent_dim = self.latent_dim
-        input_shape = (M, N, 3)
-
-        encoder_input = tfkl.Input(shape = input_shape)
-        X = Conv(32, 3)(encoder_input)
-        X = Conv(64, 3)(X)
-        X = tfkl.Flatten()(X)
-        #X = tfkl.Dense(latent_dim + latent_dim)(X)
-        Z_mu = tfkl.Dense(latent_dim)(X)
-        Z_logvar = tfkl.Dense(latent_dim, activation="relu")(X)
-        Z = Reparameterize()([Z_mu, Z_logvar])
-
-        decoder_input = tfkl.Input(shape = (latent_dim,))
-        X = tfkl.Dense(M * N * 4,  activation="relu")(decoder_input)
-        X = tfkl.Reshape(target_shape = (M//4, N//4, 64))(X)
-        X = Deconv(64, 3)(X)
-        X = Deconv(32, 3)(X)
-        decoder_output = Deconv(3, 3, strides=1, activation="sigmoid")(X)
-
-        def reconstruction_loss(X, X_pred):
-            mse = tf.losses.MeanSquaredError()
-            return mse(X, X_pred) * np.prod(input_shape)
-
-        #KL divergence between a unit normal Gaussian (the prior; p(z)) and q(z|x)
-        def kl_divergence(X, X_pred):
-            #self.C += (1/1440) # TODO use correct scalar
-            #self.C = min(self.C, 35) # TODO make variable
-            kl = -0.5 * tf.reduce_mean(1 + Z_logvar - Z_mu**2 - tf.math.exp(Z_logvar))
-            return tf.math.abs(kl)#self.gamma * tf.math.abs(kl - self.C)
-
-        def total_loss(X, X_pred):
-            return reconstruction_loss(X, X_pred) + kl_divergence(X, X_pred)
-
-        self.encoder = Model(encoder_input, [Z_mu, Z_logvar, Z])
-        self.decoder = Model(decoder_input, decoder_output)
-        self.model = Model(encoder_input, self.decoder(Z))
-        self.model.compile(optimizer=self.optimizer,
-                           loss=total_loss,
-                           metrics=[reconstruction_loss, kl_divergence])
-
-    def train(self, epochs, quiet = False):
+    def train(self, epochs, steps_for_update = None, quiet = False):
         """Train the networks in the convolutional VAE.
 
         Args:
@@ -132,39 +84,115 @@ class VAEorama(GENERICorama):
             quiet (bool): whether to give status updates
 
         """
+        if not steps_for_update:
+            steps_for_update = epochs // 10
+
         start_time = time.time()
+        for epoch in range(1, epochs + 1):
+            for train_x in self.train_dataset:
+                self.compute_apply_gradients(train_x)
 
-        # Training loop
-        for epoch in range(epochs):
-            for batch, X in enumerate(self.train_dataset):
-                loss, recon_err, kl = self.model.train_on_batch(X, X)
-                
-            self.save_model(epoch, loss, kl, recon_err)
+            if epoch % steps_for_update == 0:
+                end_time = time.time()
+                loss = tf.keras.metrics.Mean()
+                if self.test_dataset is not None:
+                    for test_x in self.test_dataset:
+                        loss(self.compute_loss(test_x))
+                elbo = -loss.result()
 
-        print("Finished training.")
+                if not quiet:
+                    print(f'Epoch: {epoch}, Test set ELBO: {elbo:.4f}, '
+                          f'time elapsed for current epoch batch {end_time - start_time:.4f}')
+                start_time = time.time()
+            if epoch == epochs:
+                break
+        return
 
-def Conv(n_filters, filter_width, strides=2, activation="relu", name=None):
-    return tfkl.Conv2D(n_filters, filter_width, 
-                       strides=strides, padding="valid",
-                       activation=activation, name=name)
+    def save_losses(self, path = "./"):
+        output = np.array([self.RECORDED_EPOCHS, self.TESTING_LOSS]).T
+        np.save(path + "epochs_testloss", output)
+        return
 
+    def save_model_weights(self, path = "/tmp/weights/"):
+        """Save the VAE network weights.
 
-def Deconv(n_filters, filter_width, strides=2, activation="relu", name=None):
-    return tfkl.Conv2DTranspose(n_filters, filter_width, 
-                                strides=strides, padding="same",
-                                activation=activation, name=name)
+        Args:
+            path (string): to directory where the network 
+            weights are saved
 
-class Reparameterize(tfkl.Layer):
+        """
+        self.model.inference_net.save_weights(
+            path + "inference_net_weights", save_format = 'tf')
+        self.model.generative_net.save_weights(
+            path + "generatives_net_weights", save_format = 'tf')
+        return
+
+    def load_model_weights(self, path = "/tmp/weights/"):
+        """Load the VAE network weights.
+
+        Args:
+            path (string): to directory where the network 
+            weights are saved
+
+        """
+        self.model.inference_net.load_weights(
+            path + "inference_net_weights")
+        self.model.generative_net.load_weights(
+            path + "generatives_net_weights")
+        return
+
+class _CVAE(tf.keras.Model):
+    """A convolutional variational autoencoder used to
+    create panoramic images.
+
+    Note: we assume there are 3 input (RGB) channels.
+
     """
-    Custom layer.
-     
-    Reparameterization trick, sample random latent vectors Z from 
-    the latent Gaussian distribution which has the following parameters 
+    def __init__(self, M, N, latent_dim):
+        super(_CVAE, self).__init__()        
+        self.input_dims = [M, N]
+        self.latent_dim = latent_dim
+        self.inference_net = tf.keras.Sequential([
+            tf.keras.layers.InputLayer(input_shape=(M, N, 3)), #(bs, M, N, 3)
+            tf.keras.layers.Conv2D(
+                filters=32, kernel_size=3, strides=(2, 2),
+                activation='relu', padding="valid"), #(bs, M/2, N/2, 32)
+            tf.keras.layers.Conv2D(
+                filters=64, kernel_size=3, strides=(2, 2), #(bs, M/4, N/4, 64)
+                activation='relu', padding="valid"),
+            tf.keras.layers.Flatten(), #(bs, (M/4) * (N/4) * 64)
+            #predicting mean and logvar
+            tf.keras.layers.Dense(latent_dim + latent_dim), # (bs, D * D)
+        ])
+        self.generative_net = tf.keras.Sequential([
+            tf.keras.layers.InputLayer(input_shape=(latent_dim,)), #(bs, D)
+            tf.keras.layers.Dense(units= M * N * 4, activation=tf.nn.relu), #M * N * 4 = (M/4)*(N/4)*64
+            tf.keras.layers.Reshape(target_shape=(M//4, N//4, 64)),
+            tf.keras.layers.Conv2DTranspose(
+                filters=64, kernel_size=3, strides=(2, 2),
+                padding="SAME", activation='relu'), #(bs, M/2, N/2, 64)
+            tf.keras.layers.Conv2DTranspose(
+                filters=32, kernel_size=3, strides=(2, 2),
+                padding="SAME", activation='relu'), #(bs, M, N, 32)
+            tf.keras.layers.Conv2DTranspose(
+                filters=3, kernel_size=3, strides=(1, 1), padding="SAME",
+                activation='sigmoid'), #(bs, M, N, 3)
+        ])
+        
+    @tf.function
+    def sample(self, eps=None):
+        if eps is None:
+            eps = tf.random.normal(shape=(100, self.latent_dim))
+        return self.decode(eps)
 
-    mean = Z_mu
-    logvar = Z_logvar
-    """
-    def call(self, inputs):
-        Z_mu, Z_logvar = inputs
-        return Z_mu + tf.math.exp(0.5 * Z_logvar) * tf.random.normal(tf.shape(Z_mu))
+    def encode(self, x):
+        mean, logvar = tf.split(self.inference_net(x), 
+                                num_or_size_splits=2, axis=1)
+        return mean, logvar
 
+    def reparameterize(self, mean, logvar):
+        eps = tf.random.normal(shape=mean.shape)
+        return eps * tf.exp(logvar * .5) + mean
+
+    def decode(self, z):
+        return self.generative_net(z)
